fixed connect and page.js
